Tabla_baja_Arreglos

+++ src/main/java/com/example/vSIAF/repository/BajaRepository.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BajaRepository extends JpaRepository<BajaEntity, Long> {
+public interface BajaRepository extends JpaRepository<BajaEntity, Integer> {
 }
+
